+++ test/unit/mongo_rep_admin/chk_mem_rep_lag.py
@@ -116,25 +116,26 @@ class UnitTest(unittest.TestCase):
 
         """
 
+        self.t_format = "%Y-%m-%d %H:%M:%S"
         self.mail = Mail()
         self.rep_status = {
             "set": "ReplicaSet",
             "members": [{"state": 1, "name": "server1",
                          "optimeDate": datetime.datetime.strptime(
-                             "2019-07-26 11:13:01", "%Y-%m-%d %H:%M:%S"),
+                             "2019-07-26 11:13:01", self.t_format),
                          "optime": True},
                         {"state": 2, "name": "server2",
                          "optimeDate": datetime.datetime.strptime(
-                             "2019-07-26 11:13:02", "%Y-%m-%d %H:%M:%S"),
+                             "2019-07-26 11:13:02", self.t_format),
                          "optime": True}]}
         self.rep_status2 = {
             "set": "ReplicaSet",
             "members": [{"state": 1, "name": "server1",
                          "optimeDate": datetime.datetime.strptime(
-                             "2019-07-26 11:13:01", "%Y-%m-%d %H:%M:%S")},
+                             "2019-07-26 11:13:01", self.t_format)},
                         {"state": 2, "name": "server2",
                          "optimeDate": datetime.datetime.strptime(
-                             "2019-07-26 11:13:02", "%Y-%m-%d %H:%M:%S")}]}
+                             "2019-07-26 11:13:02", self.t_format)}]}
         self.get_master = {"name": "master_server"}
         self.args_array = {"-z": True}
         self.args_array2 = {}